tts/tts-python-quickstart/pyproject.toml
[project]
name = "tts-python-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiofiles>=24.1.0",
  "hume[microphone]>=0.13.5",
  "python-dotenv>=1.0.1",
  "websockets>=12.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
# Show verbose output with test names and descriptions
addopts = "-v --tb=short"
# Show test docstrings in output
testpaths = ["."]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
